project_01_customer_support/requirements.txt
# Core LangChain
langchain
langchain-core
langchain-google-genai

# Vector Database
pinecone-client
langchain-pinecone

# Document Processing
pypdf
python-docx
unstructured

# Environment Management
python-dotenv

# Utilities
colorama
